refactor: ♻️ combine clients classes to the one module
Preparation for: h2oai/cloud-discovery#170
