+++ src/h2o_discovery/_load.py
@@ -2,7 +2,6 @@ import types
 from typing import Iterable
 from typing import Mapping
 
-from h2o_discovery import _async_client
 from h2o_discovery import _client
 from h2o_discovery import model
 
@@ -16,7 +15,7 @@ def load_discovery(cl: _client.Client) -> model.Discovery:
     return model.Discovery(environment=environment, services=services, clients=clients)
 
 
-async def load_discovery_async(cl: _async_client.AsyncClient) -> model.Discovery:
+async def load_discovery_async(cl: _client.AsyncClient) -> model.Discovery:
     """Loads the discovery records from the Discovery Service."""
     environment = await cl.get_environment()
     services = _get_service_map(await cl.list_services())
