+++ tests/test_client.py
@@ -2,7 +2,6 @@ import httpx
 import pytest
 import respx
 
-from h2o_discovery import _async_client
 from h2o_discovery import _client
 from h2o_discovery import model
 
@@ -46,7 +45,7 @@ async def test_async_client_get_environment_internal():
     route = respx.get("http://test.example.com:1234/v1/environment").respond(
         json=ENVIRONMENT_JSON
     )
-    cl = _async_client.AsyncClient("http://test.example.com:1234")
+    cl = _client.AsyncClient("http://test.example.com:1234")
 
     # When
     env = await cl.get_environment()
@@ -63,7 +62,7 @@ async def test_async_client_get_environment_public():
     route = respx.get(
         "https://test.example.com/.ai.h2o.cloud.discovery/v1/environment"
     ).respond(json=ENVIRONMENT_JSON)
-    cl = _async_client.AsyncClient("https://test.example.com/.ai.h2o.cloud.discovery")
+    cl = _client.AsyncClient("https://test.example.com/.ai.h2o.cloud.discovery")
 
     # When
     env = await cl.get_environment()
@@ -112,7 +111,7 @@ async def test_async_client_list_services_internal():
     route = respx.get("http://test.example.com:1234/v1/services")
     route.side_effect = SERVICES_RESPONSES
 
-    cl = _async_client.AsyncClient("http://test.example.com:1234")
+    cl = _client.AsyncClient("http://test.example.com:1234")
 
     # When
     services = await cl.list_services()
@@ -129,7 +128,7 @@ async def test_async_client_list_services_pubclic():
     route = respx.get("https://test.example.com/.ai.h2o.cloud.discovery/v1/services")
     route.side_effect = SERVICES_RESPONSES
 
-    cl = _async_client.AsyncClient("https://test.example.com/.ai.h2o.cloud.discovery")
+    cl = _client.AsyncClient("https://test.example.com/.ai.h2o.cloud.discovery")
 
     # When
     services = await cl.list_services()
@@ -180,7 +179,7 @@ async def test_async_client_list_clients_internal():
     route = respx.get("https://test.example.com:1234/v1/clients")
     route.side_effect = CLIENTS_RESPONSES
 
-    cl = _async_client.AsyncClient("https://test.example.com:1234")
+    cl = _client.AsyncClient("https://test.example.com:1234")
 
     # When
     clients = await cl.list_clients()
@@ -198,7 +197,7 @@ async def test_async_client_list_clients_public():
     route = respx.get("https://test.example.com/.ai.h2o.cloud.discovery/v1/clients")
     route.side_effect = CLIENTS_RESPONSES
 
-    cl = _async_client.AsyncClient("https://test.example.com/.ai.h2o.cloud.discovery")
+    cl = _client.AsyncClient("https://test.example.com/.ai.h2o.cloud.discovery")
 
     # When
     clients = await cl.list_clients()
@@ -227,7 +226,7 @@ def test_client_list_services_can_handle_empty_response():
 async def test_async_client_list_services_can_handle_empty_response():
     # Given
     respx.get("https://test.example.com/v1/services").respond(json={})
-    cl = _async_client.AsyncClient("https://test.example.com")
+    cl = _client.AsyncClient("https://test.example.com")
 
     # When
     services = await cl.list_services()
@@ -254,7 +253,7 @@ def test_client_list_clients_can_handle_empty_response():
 async def test_async_client_list_clients_can_handle_empty_response():
     # Given
     respx.get("https://test.example.com/v1/clients").respond(json={})
-    cl = _async_client.AsyncClient("https://test.example.com")
+    cl = _client.AsyncClient("https://test.example.com")
 
     # When
     services = await cl.list_clients()
